requirements.txt
jinja2
PyGithub
python-gitlab
